Simplify VPNProviderServerView components
Merge former ContainerView into main view and define
platform-specific subviews:

- ContainerView (composition of content + filters)
- ContentView
- FiltersView
